Convert a grid of QuadKeys to square polygons

The main argument of this function, the grid of QuadKeys points representing lat/long WG84 coordinates specifically indicate the upper-left corner of the QuadKey. To transform these coordinates into square polygons, the function supplements the grid by adding a row and column of tiles. These points introduce QuadKeys located at the border of the area (complete_grid_for_polygons). The function constructs the polygons using all the points of the grid. Note that it's possible to associate each QuadKey with its square polygon.

Usage

grid_to_polygon(data)

Arguments

data

A spatial dataset (sf) with a quadkey and POINT geometry column.

Value

A spatial dataset (sf) with the quadkey and POLYGON column.

See also

Examples

grid  <-  create_qk_grid(xmin = -59,
                      xmax = -57,
                      ymin = -35,
                      ymax = -34,
                      level = 11)

                      grid_coords <- extract_qk_coord(data = grid$data)
                      grid_coords
#> Simple feature collection with 96 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -59.0625 ymin: -34.88593 xmax: -57.12891 ymax: -33.87042
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    tileX tileY     quadkey                    geometry
#> 1    688  1236 21032130200  POINT (-59.0625 -34.88593)
#> 2    688  1235 21032130022  POINT (-59.0625 -34.74161)
#> 3    688  1234 21032130020  POINT (-59.0625 -34.59704)
#> 4    688  1233 21032130002  POINT (-59.0625 -34.45222)
#> 5    688  1232 21032130000  POINT (-59.0625 -34.30714)
#> 6    688  1231 21032112222  POINT (-59.0625 -34.16182)
#> 7    688  1230 21032112220  POINT (-59.0625 -34.01624)
#> 8    688  1229 21032112202  POINT (-59.0625 -33.87042)
#> 9    689  1236 21032130201 POINT (-58.88672 -34.88593)
#> 10   689  1235 21032130023 POINT (-58.88672 -34.74161)

                      polygrid  <-  grid_to_polygon(grid_coords)
                      polygrid
#> Simple feature collection with 96 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -59.0625 ymin: -35.03 xmax: -56.95312 ymax: -33.87042
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>        quadkey                       geometry
#> 1  21032113213 POLYGON ((-57.12891 -34.016...
#> 2  21032113231 POLYGON ((-57.12891 -34.161...
#> 3  21032113233 POLYGON ((-57.12891 -34.307...
#> 4  21032131011 POLYGON ((-57.12891 -34.452...
#> 5  21032131013 POLYGON ((-57.12891 -34.597...
#> 6  21032131031 POLYGON ((-57.12891 -34.741...
#> 7  21032131033 POLYGON ((-57.12891 -34.885...
#> 8  21032131211 POLYGON ((-57.12891 -35.03,...
#> 9  21032113212 POLYGON ((-57.30469 -34.016...
#> 10 21032113230 POLYGON ((-57.30469 -34.161...